feat: Complete modularization of code
- Moved validateEmail, validatePassword, showError, hideError functions
  to validate.js
- Separated passwordVisibility, checkLoginButtonStatus,
  handleEmailValidation, handlePasswordValidation, handleFormSubmit
functions into login.js
- Created main script file index.js and imported login.js module
