+++ index.html
@@ -118,5 +118,6 @@
         <a href="http://www.youtube.com" target="_blank"><img class="no-drag" src="./imgs/ic_youtube.png" alt="youtube" width="20" height="20"></a>
       </div>               
     </footer>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
