+++ login.html
@@ -25,11 +25,11 @@
       <form class="login-form">
         <div class="email-container">
           <label for="user-email">이메일</label>
-          <input class="input" type="text" id="user-email" name="email" placeholder="이메일을 입력해주세요." onchange="checkLoginButtonStatus()">
+          <input class="input" type="text" id="user-email" name="email" placeholder="이메일을 입력해주세요.">
         </div>
         <div class="password-container">
           <label for="password">비밀번호</label>
-          <input class="input" type="password" id="password" name="password" placeholder="비밀번호를 입력해주세요." onchange="checkLoginButtonStatus()">
+          <input class="input" type="password" id="password" name="password" placeholder="비밀번호를 입력해주세요.">
           <img class="visibility-icon" src="./imgs/btn_visibility_off.png" alt="visible off"  width="24" height="24">
         </div>
         <button id="button">로그인</button>
@@ -42,7 +42,7 @@
         </div>
         <p class="select-status">판다마켓이 처음이신가요? <a href="/signup.html">회원가입</a></p>
       </form>
-      <script src="./index.js"></script>
+      <script type="module" src="./login.js"></script>
     </div>
   </body>
 </html>
